debate_judge/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
wikipedia>=1.4.0
colorama>=0.4.6
requests>=2.28.0
beautifulsoup4>=4.12.0
